+++ lightkurve/periodogram.py
@@ -13,10 +13,15 @@ from mpl_toolkits.axes_grid1.inset_locator import inset_axes
 
 import astropy
 from astropy.table import Table
-from astropy.stats import LombScargle
 from astropy import units as u
 from astropy.units import cds
 from astropy.convolution import convolve, Box1DKernel, Gaussian1DKernel
+
+# LombScargle was moved from astropy.stats to astropy.timeseries in AstroPy v3.2
+try:
+    from astropy.timeseries import LombScargle
+except ImportError:
+    from astropy.stats import LombScargle
 
 from scipy.optimize import curve_fit
 from scipy.signal import find_peaks
@@ -529,7 +534,7 @@ class Periodogram(object):
             Helper object to run asteroseismology methods
         """
         from .seismology import SeismologyButler
-        return SeismologyButler.from_periodogram(periodogram=self, **kwargs)
+        return SeismologyButler(self)
 
 
 class SNRPeriodogram(Periodogram):
@@ -570,7 +575,6 @@ class LombScarglePeriodogram(Periodogram):
     def __init__(self, *args, **kwargs):
         self._LS_object = kwargs.pop("ls_obj", None)
         self.nterms = kwargs.pop("nterms", 1)
-        self.lc = kwargs.pop("lc", None)
         self.ls_method = kwargs.pop("ls_method", 'fastchi2')
         super(LombScarglePeriodogram, self).__init__(*args, **kwargs)
 
@@ -708,18 +712,6 @@ class LombScarglePeriodogram(Periodogram):
         Periodogram : `Periodogram` object
             Returns a Periodogram object extracted from the lightcurve.
         """
-        # If the defaults are used, issue a warning to point out they changed!
-        # if normalization == 'amplitude' and freq_unit is None and oversample_factor is None:
-        #     warnings.warn("As of Lightkurve v1.0.0 (Apr 2019), the default behavior "
-        #                   "of Lomb Scargle periodograms changed to use "
-        #                   "normalization='amplitude' and oversample_factor=5 "
-        #                   "(the previous defaults were normalization='psd' and "
-        #                   "oversample_factor=1). You can suppress this warning using "
-        #                   "`warnings.filterwarnings('ignore', category=lk.LightkurveWarning)`.",
-        #                   LightkurveWarning)
-
-        # Carry the input light curve
-
         # Input validation for spectrum type
         if normalization not in ('psd', 'amplitude'):
             raise ValueError("The `normalization` parameter must be one of "
@@ -834,15 +826,12 @@ class LombScarglePeriodogram(Periodogram):
         # Convert to desired units
         frequency = u.Quantity(frequency, freq_unit)
 
-
-
         if period is not None:
             if ls_method is 'fastchi2':
                 ls_method = 'chi2'
             elif ls_method is 'fast':
                 ls_method = 'slow'
 
-#            ls_method = 'chi2'
             log.warning("You have passed an evenly-spaced grid of periods. "
                         "These are not evenly spaced in frequency space.\n"
                         "Method has been set to '{}' to allow for this.".format(ls_method))
@@ -854,7 +843,6 @@ class LombScarglePeriodogram(Periodogram):
                             "Please refer to the `astropy.timeseries.periodogram.LombScargle` documentation.".format(ls_method),
                           LightkurveWarning)
             nterms = 1
-
 
         flux_scaling = 1e6
         if float(astropy.__version__[0]) >= 3:
@@ -884,17 +872,15 @@ class LombScarglePeriodogram(Periodogram):
         return LombScarglePeriodogram(frequency=frequency, power=power, nyquist=nyquist,
                                       targetid=lc.targetid, label=lc.label,
                                       default_view=default_view, ls_obj=LS,
-                                      nterms=nterms, ls_method=ls_method, lc=lc)
+                                      nterms=nterms, ls_method=ls_method)
 
-
-    def flux_model(self, time=None, frequency=None):
-        '''Obtain the flux model for a given frequency and time
+    def model(self, time, frequency=None):
+        """Obtain the flux model for a given frequency and time
 
         Parameters
         ----------
         time : np.ndarray
-            Time points to evaluate model. Default is the original
-            input light curve time.
+            Time points to evaluate model.
         frequency : frequency to evaluate model. Default is the frequency at
                     max power.
 
@@ -902,21 +888,14 @@ class LombScarglePeriodogram(Periodogram):
         -------
         result : lightkurve.LightCurve
             Model object with the time and flux model
-        '''
+        """
         if self._LS_object is None:
             raise ValueError('No `astropy` Lomb Scargle object exists.')
-
-        if time is None:
-            if self.lc.time_format in ['bkjd', 'btjd', 'd', 'days', 'day', None]:
-                time = self.lc.time.copy() * u.day
-            else:
-                raise NotImplementedError('time in format {} is not supported.'.format(self.lc.time_format))
         if frequency is None:
             frequency = self.frequency_at_max_power
         f = self._LS_object.model(time, frequency)
         return LightCurve(time, f, label='LS Model', meta={'frequency':frequency},
-                            targetid='{} LS Model'.format(self.lc.targetid)).normalize()
-
+                            targetid='{} LS Model'.format(self.targetid)).normalize()
 
 
 class BoxLeastSquaresPeriodogram(Periodogram):
@@ -942,10 +921,16 @@ class BoxLeastSquaresPeriodogram(Periodogram):
     @staticmethod
     def from_lightcurve(lc, **kwargs):
         """Creates a Periodogram from a LightCurve using the Box Least Squares (BLS) method."""
+        # BoxLeastSquares was added to `astropy.stats` in AstroPy v3.1 and then
+        # moved to `astropy.timeseries` in v3.2, which makes the import below
+        # somewhat complicated.
         try:
-            from astropy.stats import BoxLeastSquares
+            from astropy.timeseries import BoxLeastSquares
         except ImportError:
-            raise ImportError("BLS requires AstroPy v3.1 or later")
+            try:
+                from astropy.stats import BoxLeastSquares
+            except ImportError:
+                raise ImportError("BLS requires AstroPy v3.1 or later")
 
         # Validate user input for `lc`
         # (BoxLeastSquares will not work if flux or flux_err contain NaNs)