+++ lightkurve/seismology/__init__.py
@@ -1,0 +1,2 @@
+"""This sub-package provides classes and functions for asteroseismology."""
+from .butler import *